+++ pytsite/admin/_api.py
@@ -1,7 +1,8 @@
 """Admin API Functions
 """
 from pytsite import tpl as _tpl, widget as _widget, core_version_str as _version_str, core_url as _core_url, \
-    core_name as _core_name, browser as _browser, assetman as _assetman
+    core_name as _core_name, browser as _browser
+from . import _sidebar
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -11,14 +12,8 @@ __license__ = 'MIT'
 def render(content: str) -> str:
     """Render admin page with content.
     """
-    _browser.include('bootstrap') 
-    _browser.include('font-awesome')
-    _assetman.add('pytsite.admin@AdminLTE/css/AdminLTE.min.css')
-    _assetman.add('pytsite.admin@AdminLTE/css/skins/skin-blue.min.css')
-    _assetman.add('pytsite.admin@css/custom.css')
-    _assetman.add('pytsite.admin@AdminLTE/js/app.js')
-    
     return _tpl.render('pytsite.admin@html', {
+        'admin_sidebar': _sidebar.render(),
         'admin_language_nav': _widget.select.LanguageNav('admin-language-nav', dropdown=True),
         'content': content,
         'core_name': _core_name,