+++ pytsite/assetman/_api.py
@@ -8,16 +8,19 @@ __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-
 _packages = {}
+
 _locations = {}
 _last_weight = {}
-_last_permanent_weight = {}
-_last_inline_weight = {}
+
+_p_locations = {}
+_last_p_weight = 0
+
 _inline = {}
+_last_i_weight = {}
 
 
-def register_package(package_name: str, assets_dir: str='res/assets'):
+def register_package(package_name: str, assets_dir: str = 'res/assets'):
     """Register assets container.
     """
     spec = _find_spec(package_name)
@@ -28,27 +31,19 @@ def register_package(package_name: str, assets_dir: str='res/assets'):
     if not _path.isdir(dir_path):
         FileNotFoundError("Directory '{}' is not found.".format(dir_path))
 
-    tid = _threading.get_id()
-    if tid not in _packages:
-        _packages[tid] = {}
-
-    if package_name in _packages[tid]:
+    if package_name in _packages:
         raise RuntimeError("Package '{}' is already registered.".format(package_name))
 
-    _packages[tid][package_name] = dir_path
+    _packages[package_name] = dir_path
 
 
 def get_packages() -> dict:
     """Get registered packages.
     """
-    tid = _threading.get_id()
-    if tid not in _packages:
-        _packages[tid] = {}
-
-    return _packages[tid]
+    return _packages
 
 
-def add(location: str, collection: str=None, weight=0, permanent=False):
+def add(location: str, collection: str = None, weight: int = 0, permanent=False, path_prefix: str = None):
     """Add an asset.
     """
     # Determine collection
@@ -60,40 +55,47 @@ def add(location: str, collection: str=None, weight=0, permanent=False):
         else:
             raise ValueError("Cannot determine collection of location '{}'.".format(location))
 
+    if path_prefix and not path_prefix.startswith('/'):
+        path_prefix = '/' + path_prefix
+
     tid = _threading.get_id()
     if tid not in _locations:
         _locations[tid] = {}
-        _last_weight[tid] = 0
 
-    # Add only if it isn't added yet
-    if location not in _locations[tid]:
-        # Determine weight
-        if not weight:
-            _last_weight[tid] += 10
-            weight = _last_weight[tid]
+    location_hash = hash((location, path_prefix))
 
-        elif weight > _last_weight[tid]:
-            _last_weight[tid] = weight
-
+    if location_hash not in _p_locations and location_hash not in _locations[tid]:
         if permanent:
-            _last_permanent_weight[tid] = weight
+            global _last_p_weight
 
-        _locations[tid][location] = (collection, weight, permanent)
+            if not weight:
+                _last_p_weight += 10
+                weight = _last_p_weight
+            elif weight > _last_p_weight:
+                _last_p_weight = weight
+
+            _p_locations[location_hash] = (location, collection, weight, path_prefix)
+        else:
+            if not weight:
+                _last_weight[tid] += 10
+                weight = _last_weight[tid]
+            elif weight > _last_weight[tid]:
+                _last_weight[tid] = weight
+
+            _locations[tid][location_hash] = (location, collection, weight, path_prefix)
 
 
-def add_inline(s: str, weight=0, forever=False):
+def add_inline(s: str, weight=0):
     """Add a code which intended to output in the document.
     """
     tid = _threading.get_id()
-    if tid not in _inline:
-        _inline[tid] = []
 
     if not weight:
-        _last_inline_weight[tid] += 10
-    elif weight > _last_inline_weight[tid]:
-        _last_inline_weight[tid] = weight
+        _last_i_weight[tid] += 10
+    elif weight > _last_i_weight[tid]:
+        _last_i_weight[tid] = weight
 
-    _inline[tid].append((s, weight, forever))
+    _inline[tid].append((s, weight))
 
 
 def remove(location):
@@ -103,13 +105,12 @@ def remove(location):
     if tid not in _locations:
         return
 
+    # Location as a string
     if isinstance(location, str):
-        _locations[tid] = {k: v for k, v in _locations[tid] if k != location}
-
-    # Compiled regular expression
+        _locations[tid] = {k: v for k, v in _locations[tid].items() if location != v[0]}
+    # Location as a compiled regular expression
     elif not isinstance(location, str) and location.__class__.__name__ == 'SRE_Pattern':
-        _locations[tid] = {k: v for k, v in _locations[tid] if not location.match(k)}
-
+        _locations[tid] = {k: v for k, v in _locations[tid].items() if not location.match(v[0])}
     else:
         raise TypeError('String or compiled regular expression expected.')
 
@@ -117,41 +118,43 @@ def remove(location):
 def reset():
     """Remove all previously added locations and inline code except 'permanent'.
     """
+    global _last_weight
+
     tid = _threading.get_id()
 
-    # Filter out all except 'permanent' items
-    if tid in _locations:
-        _locations[tid] = {k: v for k, v in _locations[tid].items() if v[2]}
-
-    # Filter out all inline code except 'permanent' items
-    if tid in _inline:
-        _inline[tid] = [s for s in _inline[tid] if s[2]]
-
-    if tid not in _last_permanent_weight:
-        _last_permanent_weight[tid] = 0
-
-    if tid not in _last_weight:
-        _last_weight[tid] = 0
-
-    # Non-permanent locations should never stand before permanent ones
-    _last_weight[tid] = _last_permanent_weight[tid]
+    _locations[tid] = {}
+    _last_weight[tid] = 0
+    _inline[tid] = []
+    _last_i_weight[tid] = 0
 
 
-def get_locations(collection: str=None) -> list:
+def get_locations(collection: str = None, filter_path: bool = True) -> tuple:
     tid = _threading.get_id()
 
-    if tid not in _locations:
-        return []
+    p_locations = _p_locations.values()
+    locations = _locations[tid].values()
 
+    locations = sorted(p_locations, key=lambda x: x[2]) + sorted(locations, key=lambda x: x[2])
+
+    # Filter by collection
     if collection:
-        # Select all locations of particular collection
-        locations = [(k, v[0], v[1], v[2]) for k, v in _locations[tid].items() if v[0] == collection]
-    else:
-        # All locations
-        locations = [(k, v[0], v[1], v[2]) for k, v in _locations[tid].items()]
+        locations = [l for l in locations if l[1] == collection]
 
-    # Sort by weight
-    return [l for l in sorted(locations, key=lambda x: x[2])]
+    # Filter by path
+    if filter_path:
+        current_path = _router.current_path()
+        locations = [l for l in locations if (l[3] is None or current_path.startswith(l[3]))]
+
+    # Build unique list.
+    # Duplicates are possible because same location may be added more than once with different path prefixes.
+    added = []
+    r = []
+    for l in locations:
+        if l[0] not in added:
+            added.append(l[0])
+            r.append(l)
+
+    return r
 
 
 def get_inline() -> list:
@@ -187,7 +190,7 @@ def dump_inline() -> str:
 
     tid = _threading.get_id()
     if tid in _inline:
-        for item in _inline:
+        for item in _inline[tid]:
             r += item[0]
 
     return r
@@ -204,19 +207,15 @@ def url(location: str) -> str:
     return _router.url('/assets/{}/{}'.format(package_name, asset_path), strip_lang=True)
 
 
-def get_urls(collection: str=None) -> list:
+def get_urls(collection: str = None, filter_path: bool = True) -> list:
     """Get URLs of all locations in the collection.
     """
-    return [url(l[0]) for l in get_locations(collection)]
+    return [url(l[0]) for l in get_locations(collection, filter_path)]
 
 
 def _split_asset_location_info(location: str) -> dict:
     """Split asset path into package name and asset path.
     """
-    tid = _threading.get_id()
-    if tid not in _packages:
-        _packages[tid] = {}
-
     package_name = 'app'
     asset_path = location
     path_parts = location.split('@')
@@ -224,7 +223,7 @@ def _split_asset_location_info(location: str) -> dict:
         package_name = path_parts[0]
         asset_path = path_parts[1]
 
-    if package_name not in _packages[tid]:
+    if package_name not in _packages:
         raise RuntimeError("Package '{}' is not registered.".format(package_name))
 
     return package_name, asset_path