+++ pytsite/google/__init__.py
@@ -16,7 +16,7 @@ __maps_url_re = _re.compile('^https://maps.googleapis.com/maps/api/js')
 __libraries_re = _re.compile('libraries=([a-z]+)')
 
 
-def __browser_library_maps_callback(permanent: bool, **kwargs):
+def __browser_library_maps_callback(permanent: bool, path_prefix: str, **kwargs):
     # Google Map API key is required
     api_key = _reg.get('google.maps.client_key')
     if not api_key:
@@ -29,20 +29,20 @@ def __browser_library_maps_callback(permanent: bool, **kwargs):
     for loc in _assetman.get_locations('js'):
         if __maps_url_re.match(loc[0]) and 'libraries=' in loc[0]:
             # Remember weight of the location to place new one at the same position
-            asset_weight = loc[1]
+            asset_weight = loc[2]
 
             # Extracting Google libraries list
             libs += __libraries_re.findall(loc[0])[0].split(',')
 
-    # Remove probably previous added link too Google Maps API JS
-    _assetman.remove(__maps_url_re, 'js')
+    # Remove probably previously added link too Google Maps API JS
+    _assetman.remove(__maps_url_re)
 
     # Add assets
     libs = ','.join(libs)
     js = 'https://maps.googleapis.com/maps/api/js?libraries={}&amp;language={}&amp;key={}'\
         .format(libs, _lang.get_current(), api_key)
 
-    _assetman.add(js, 'js', asset_weight, permanent)
+    _assetman.add(js, 'js', asset_weight, permanent, path_prefix)
 
 
 def __init():