+++ pytsite/metatag/__init__.py
@@ -1,6 +1,6 @@
 """PytSite MetaTag Module.
 """
-from ._functions import dump, dump_all, get, reset, t_set
+from ._api import dump, dump_all, get, reset, t_set
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'